refactor: move cache to utils folder
