+++ aio/aio-proxy/aio_proxy/utils/cache.py
@@ -3,7 +3,7 @@ import logging
 from collections.abc import Callable
 
 from aio_proxy.response.helpers import hash_string
-from aio_proxy.search.cache.redis import RedisClient
+from aio_proxy.utils.redis import RedisClient
 
 
 def build_key(key):
